+++ src/Plivo/Resources/Message/MessageInterface.php
@@ -9,6 +9,7 @@ use Plivo\Exceptions\PlivoRestException;
 use Plivo\Exceptions\PlivoResponseException;
 use Plivo\Util\ArrayOperations;
 use Plivo\Util\Template;
+use Plivo\Util\Interactive;
 
 use Plivo\MessageClient;
 use Plivo\Resources\ResourceInterface;
@@ -155,6 +156,7 @@ class MessageInterface extends ResourceInterface
             $powerpackUUID = isset($optionalArgs['powerpackUUID']) ? $optionalArgs['powerpackUUID'] : null;
         }
         $template = isset($optionalArgs['template']) ? $optionalArgs['template'] : null;       
+        $interactive = isset($optionalArgs['interactive']) ? $optionalArgs['interactive'] : null;       
         if (is_array($dst)){
             $mandatoryArgs = [
                 'dst' => implode('<', $dst),
@@ -195,6 +197,11 @@ class MessageInterface extends ResourceInterface
                 'Template paramater is only applicable when message_type is whatsapp'
             );
         }
+        if (isset($optionalArgs['type'])  && $optionalArgs['type'] != 'whatsapp' && !is_null($interactive)){
+            throw new PlivoValidationException(
+                'Interactive paramater is only applicable when message_type is whatsapp'
+            );
+        }
 
         if(!is_null($template)){
             $err = Template::validateTemplate($template);
@@ -207,6 +214,17 @@ class MessageInterface extends ResourceInterface
             $optionalArgs['template'] = json_decode($template,True);
         }
         
+        if(!is_null($interactive)){
+            $err = Interactive::validateInteractive($interactive);
+            if (!is_null($err))
+            {
+                throw new PlivoValidationException(
+                    $err
+                );
+            }
+            $optionalArgs['interactive'] = json_decode($interactive,True);
+        }
+
         $response = $this->client->update(
             $this->uri,
             array_merge($mandatoryArgs,  $optionalArgs, ['src' => $src, 'powerpack_uuid' => $powerpackUUID, 'text' => $text])